requirements/local.txt
-r ./base.txt

# Debugging
ipdb==0.13.9

# Tools
django-extensions==3.1.3

# Testing
mypy==0.910
mypy-extensions==0.4.3
pytest==6.2.4
pytest-django==4.4.0
pytest-sugar==0.9.4
factory-boy==3.2.0

# Code quality
flake8==3.9.2
